Updated annotation & utils
